# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Security scanning platform — manages targets, scan jobs, reports, and triage rules.",
        "title": "NexusSec API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/scans": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lấy danh sách tất cả scan job của user hiện tại (tối đa 50, sắp xếp mới nhất).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scans"
                ],
                "summary": "List scan jobs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/internal_gateway_handler.scanListItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Tạo scan job mới (status=pending) và đẩy vào RabbitMQ để xử lý bất đồng bộ. Trả về 202 Accepted.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scans"
                ],
                "summary": "Create a scan job",
                "parameters": [
                    {
                        "description": "Target ID và loại scan",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_handler.createScanRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Scan job đã được chấp nhận",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_gateway_handler.scanJobResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ hoặc SSRF blocked",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Target không tồn tại",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "429": {
                        "description": "Quá nhiều request",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    }
                }
            }
        },
        "/scans/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lấy chi tiết một scan job theo ID. User phải là chủ sở hữu.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scans"
                ],
                "summary": "Get scan job detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Scan Job ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_gateway_handler.scanListItem"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Scan không tồn tại",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_nexussec_nexussec_pkg_response.JSONResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "internal_gateway_handler.createScanRequest": {
            "type": "object",
            "required": [
                "scan_type",
                "target_id"
            ],
            "properties": {
                "scan_type": {
                    "type": "string",
                    "enum": [
                        "zap",
                        "nmap",
                        "full"
                    ]
                },
                "target_id": {
                    "type": "string"
                }
            }
        },
...
        "internal_gateway_handler.scanJobResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "progress": {
                    "type": "integer"
                },
                "scan_type": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "target_url": {
                    "type": "string"
                }
            }
        },
        "internal_gateway_handler.scanListItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "progress": {
                    "type": "integer"
                },
                "scan_type": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "target_url": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Nhập: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}